added: staff order page new functional

+++ src/main/java/model/Entities/Driver.java
@@ -21,7 +21,7 @@ public class Driver {
     @Column(nullable = false)
     private String lastName;
 
-    private double thisMonthHours;
+    private float thisMonthHours;
 
     @ManyToOne
     @Cascade(CascadeType.PERSIST)
@@ -67,7 +67,7 @@ public class Driver {
      * @param currentWagon   the current wagon
      * @param currentOrder   the current order
      */
-    public Driver(String firstName, String lastName, double thisMonthHours, DriverStatus driverStatus, City currentCity, Wagon currentWagon, Order currentOrder) {
+    public Driver(String firstName, String lastName, float thisMonthHours, DriverStatus driverStatus, City currentCity, Wagon currentWagon, Order currentOrder) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.thisMonthHours = thisMonthHours;
@@ -156,7 +156,7 @@ public class Driver {
      *
      * @param thisMonthHours the this month hours
      */
-    public void setThisMonthHours(double thisMonthHours) {
+    public void setThisMonthHours(float thisMonthHours) {
         this.thisMonthHours = thisMonthHours;
     }
